Dungeon씬 보스 수정, Desert씬 잡몹 추가
사운드만 추가 필요

+++ Assets/Scripts/Monster/TrashMob/TrashMob.cs
@@ -224,7 +224,7 @@ public class TrashMob : MonoBehaviour, IHitable
         if (!invincible)
         {
             GameObject damageUI = PoolManager.Instance.Get("DamageFontUI");
-            damageUI.GetComponent<DamageUI>().GetDamageFont(transform.position, (int)damage);
+            damageUI.GetComponent<DamageUI>().GetDamageFont(transform.position, damage);
 
             hpBarUI.SetActive(true);
